pyproject.toml
[tool.poetry]
name = "Streaming Processing Pipeline"
version = "0.1.0"
description = ""
authors = ["Joao Pedro <jpfdsilva@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
apache-flink = "1.18"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
